refactor: Move shape definitions to a dedicated file
This commit refactors the theme structure by moving all shape-related definitions out of `Theme.kt` and into a new, dedicated `Shapes.kt` file. This improves code organization and modularity within the `ui.theme` package.

Key changes:
- Created `app/src/main/java/com/amrubio27/compose_superheroes/ui/theme/Shapes.kt`.
- Moved the `AppShapes` and `HeroImageShape` definitions from `Theme.kt` to the new `Shapes.kt` file.
- Removed the now-redundant shape definitions and related imports from `Theme.kt`.
