+++ app/src/main/java/com/amrubio27/compose_superheroes/ui/theme/Shapes.kt
@@ -1,0 +1,28 @@
+package com.amrubio27.compose_superheroes.ui.theme
+
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.material3.Shapes
+import androidx.compose.ui.unit.dp
+
+/**
+ * Shapes personalizados de la aplicación.
+ * Alineados con Material3 pero con valores específicos del diseño.
+ */
+val AppShapes = Shapes(
+    extraSmall = RoundedCornerShape(4.dp),
+    small = RoundedCornerShape(8.dp),
+    medium = RoundedCornerShape(12.dp),
+    large = RoundedCornerShape(16.dp),
+    extraLarge = RoundedCornerShape(24.dp)
+)
+
+/**
+ * Shape especial para las imágenes de cabecera de héroe.
+ * Solo redondea las esquinas inferiores.
+ */
+val HeroImageShape = RoundedCornerShape(
+    topStart = 0.dp,
+    topEnd = 0.dp,
+    bottomStart = 24.dp,
+    bottomEnd = 24.dp
+)
